requirement.txt
beautifulsoup4
requests
google
Flask
openai
python-dotenv